+++ STRUCTURE.md
@@ -4,26 +4,37 @@
 <repo-root>
 ├── steerable_motion.py
 ├── headless.py
-├── sm_functions/
+├── source/
 │   ├── __init__.py
 │   ├── common_utils.py
-│   ├── travel_between_images.py
-│   └── different_pose.py
+│   ├── db_operations.py
+│   ├── specialized_handlers.py
+│   ├── video_utils.py
+│   └── sm_functions/
+│       ├── __init__.py
+│       ├── travel_between_images.py
+│       └── different_pose.py
 ├── Wan2GP/  ← third-party video-generation engine (keep high-level only)
 └── STRUCTURE.md  (this file)
 ```
 
 ## Top-level scripts
 
-**steerable_motion.py** – CLI orchestrator. Parses user arguments for the two primary tasks, ensures the SQLite `tasks` table exists, sets global/debug flags and then delegates to task handlers in `sm_functions`.
-
 **headless.py** – Headless server that continuously polls the `tasks` database, claims work, and drives the Wan2GP generator (`wgp.py`). Includes extra handlers for OpenPose and RIFE interpolation tasks and can upload outputs to Supabase storage.
 
-## sm_functions/ package
+## source/ package
 
-A light-weight, testable wrapper around the bulky `steerable_motion.py` logic.  It holds:
+This is the main application package.
 
-* **common_utils.py** – Reusable helpers (DB access, file downloads, ffmpeg helpers, MediaPipe keypoint interpolation, debug utilities, etc.)
+* **common_utils.py** – Reusable helpers (file downloads, ffmpeg helpers, MediaPipe keypoint interpolation, debug utilities, etc.)
+* **db_operations.py** – Handles all database interactions for both SQLite and Supabase.
+* **specialized_handlers.py** - Contains handlers for specific, non-standard tasks like OpenPose generation and RIFE interpolation.
+* **video_utils.py** - Provides utilities for video manipulation like cross-fading, frame extraction, and color matching.
+
+### source/sm_functions/ sub-package
+
+A light-weight, testable wrapper around the bulky `steerable_motion.py` logic.  It holds task-specific handlers:
+
 * **travel_between_images.py** – Implements the segment-by-segment interpolation pipeline between multiple anchor images.  Builds guide videos, queues generation tasks, stitches outputs.
 * **different_pose.py** – Generates a new pose for a single image using an OpenPose-driven guide video plus optional RIFE interpolation for smoothness.
 * **__init__.py** – Re-exports public APIs (`run_travel_between_images_task`, `run_different_pose_task`) and common utilities for convenient importing.
