+++ source/sm_functions/different_pose.py
@@ -9,7 +9,7 @@ import cv2 # pip install opencv-python
 from PIL import Image # pip install Pillow
 
 # Import from the new common_utils module
-from .common_utils import (
+from ..common_utils import (
     DEBUG_MODE, dprint, generate_unique_task_id, add_task_to_db, poll_task_status,
     save_frame_from_video, create_pose_interpolated_guide_video,
     generate_different_pose_debug_video_summary # For debug mode
